Removed tailwindcss for loadtime performance improvements
