+++ .gitignore
@@ -1,29 +1,5 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
-
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-
-# testing
-/coverage
-
-# production
-/build
-
-# misc
-.DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# firebase
-.firebase/
-database-debug.log
-firebase-debug.log
+node_modules
+.firebase
 ui-debug.log
+database-debug.log
+build
