+++ package.json
@@ -3,15 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^6.1.1",
-    "@tailwindcss/forms": "^0.2.1",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.7.1",
     "firebase": "^8.2.8",
     "firebaseui": "^4.7.3",
     "immer": "^8.0.1",
-    "postcss": "^7.0.32",
     "react": "^17.0.1",
     "react-colorful": "^5.0.0",
     "react-dom": "^17.0.1",
@@ -21,9 +18,9 @@
     "web-vitals": "^1.1.0"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -44,8 +41,5 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {
-    "@tailwindcss/postcss7-compat": "^2.0.3",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.0.3"
-  }
+  "devDependencies": {}
 }
