+++ src/components/Home/Dashboard.js
@@ -1,28 +1,13 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect, useReducer } from 'react'
 import { useHistory } from 'react-router-dom'
 import { HexColorPicker, HexColorInput } from "react-colorful";
-import { useAuth } from '../contexts/AuthContext'
 import 'firebase/database'
-import app from "../firebase"
+import app from "../../firebase"
+import { useAuth } from '../../contexts/AuthContext'
+import Newsfeed from "./Newsfeed"
 
 export default function Dashboard() {
     const { currentUser, logout } = useAuth()
-    const currentUserEntriesRef =  app.database().ref("entries").child(currentUser.uid)
-
-    useEffect(() => {
-        currentUserEntriesRef.on('value', (snapshot => {
-            const data = snapshot.val()
-            console.log("DATA: ")
-            console.log(data)
-            if (data) {
-                setEntries(data)
-            }
-        }))
-
-        return () => {
-            currentUserEntriesRef.off()
-        }
-    }, [])
 
     const [error, setError] = useState("")
     const [shouldShowNewMoodModal, setNewMoodModal] = useState(false)
@@ -31,18 +16,10 @@ export default function Dashboard() {
     const [color, setColor] = useState('#1e88e5');
     const [note, setNote] = useState("")
 
-    const [entries, setEntries] = useState([])
+    const currentUserEntriesRef =  app.database().ref("entries").child(currentUser.uid)
 
     function handleTextEdit(e) {
         setNote(e.target.value)
-    }
-
-    function popUpNewModal() {
-        setNewMoodModal(true)
-    }
-
-    function closeModal() {
-        setNewMoodModal(false)
     }
 
     function createNewEntry(e) {
@@ -67,17 +44,6 @@ export default function Dashboard() {
         }
     }
 
-    function formatDate(utcSeconds) {
-        let date = new Date(0)
-        console.log(utcSeconds)
-        date.setUTCMilliseconds(utcSeconds)
-        return (
-            <div className="sticky self-center bg-white rounded-full shadow-md p-2">
-                {date.toLocaleDateString("en-US")}
-            </div>
-        )
-    }
-
     return (
         <>
         { shouldShowNewMoodModal == true && <div className="fixed left-0 top-0 z-10 h-screen w-screen">
@@ -85,12 +51,12 @@ export default function Dashboard() {
             <div className="flex flex-row w-full">
                 <h2 className="text-lg font-bold h-8">New Mood</h2>
                 <button
-                    onClick={closeModal}
+                    onClick={setNewMoodModal(false)}
                     className="inline-block ml-auto w-6 h-6 text-sm rounded-full bg-indigo-600 text-white"
                 >x</button>
             </div>
-                <form onSubmit={createNewEntry} className="flex flex-col sm:flex-wrap min-w-full h-5/6">
-                    <HexColorPicker color={color} onChange={setColor} />
+                <form className="flex flex-col sm:flex-wrap min-w-full h-5/6">
+                    <HexColorPicker color={color} onChange={setColor(color)} />
                     <HexColorInput className="mt-1 border-4 border-indigo-100 rounded-md w-200px" color={color} onChange={setColor} />
                     <label className="mt-1 sm:w-1/2 h-24 sm:h-4/6">
                         <h3 className="font-bold">Notes</h3>
@@ -121,19 +87,14 @@ export default function Dashboard() {
             <div className="flex flex-col h-full">
                 <div className="flex flex-col h-48 sm:h-64 items-center justify-center">
                     <button
-                        onClick={popUpNewModal}
+                        onClick={setNewMoodModal(true)}
                         className="rounded-full h-8 w-8 flex items-center justify-center text-white bg-indigo-300 hover:bg-indigo-500 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-600"
                     >+</button>
                     New Mood
                 </div>
-                {Object.keys(entries).reverse().map((key, index) => {
-                    return (
-                        <div className="flex flex-col items-end h-color shadow-md" style={{backgroundColor: entries[key].color}} >
-                            {formatDate(key)}
-                            <button className="w-12">*</button>
-                        </div>
-                    )
-                })}
+                {/* <Newsfeed 
+                    dbRef={currentUserEntriesRef}
+                /> */}
             </div>
         </div>
         </>
