+++ src/index.css
@@ -1,8 +1,3 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-
 .h-color {
   height: 45vh;
 }
@@ -16,5 +11,5 @@
 }
 
 body {
-  @apply font-mono text-indigo-900
+  color: indigo;
 }
